requirements.txt
streamlit
pdfplumber
openai
sqlalchemy
pandas
python-dotenv
pypdf2
reportlab
